requirements.txt
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.10.3
lxml==4.4.2
python-docx==0.8.10